Revert "feat(exporter): split exporter bundle with ESM code splitting and laz…"
This reverts commit eaaf9cf94d081cafda3d8da8e385ce5b7d871e1c.
