+++ frontend/build.mjs
@@ -79,10 +79,10 @@ await buildInParallel(
         },
         {
             name: 'Exporter',
-            entryPoints: { exporter: 'src/exporter/index.tsx' },
-            splitting: true,
-            format: 'esm',
-            outdir: path.resolve(__dirname, 'dist'),
+            globalName: 'posthogExporter',
+            entryPoints: ['src/exporter/index.tsx'],
+            format: 'iife',
+            outfile: path.resolve(__dirname, 'dist', 'exporter.js'),
             ...common,
         },
         {
