--- frontend/src/scenes/dashboard/dashboardConstants.ts
@@ -1,2 +1,0 @@
-export const AUTO_REFRESH_INITIAL_INTERVAL_SECONDS = 1800
-export const SHARED_DASHBOARD_AUTO_FORCE_IF_STALE_MINUTES = AUTO_REFRESH_INITIAL_INTERVAL_SECONDS / 60
